Add files via upload
This project analyzes data from the CDC related to the distribution of COVID-19 vaccines. It aims to investigate disparities in vaccine
administration and distribution, with a particular focus on geographic and age-based variations, especially among the population
aged 65 and older. The ratio of Admin_per_100k / Dist_per_100k effectively measures how well the distributed vaccines are being
utilized.